+++ include/freerdp/cache/glyph.h
@@ -34,15 +34,15 @@ typedef struct rdp_glyph_cache rdpGlyphCache;
 
 struct _GLYPH_CACHE
 {
-	uint16 number;
-	uint16 maxCellSize;
+	uint32 number;
+	uint32 maxCellSize;
 	rdpGlyph** entries;
 };
 
 struct _FRAGMENT_CACHE_ENTRY
 {
 	void* fragment;
-	uint16 size;
+	uint32 size;
 };
 
 struct _FRAGMENT_CACHE
@@ -58,11 +58,11 @@ struct rdp_glyph_cache
 	FRAGMENT_CACHE fragCache;
 };
 
-FREERDP_API rdpGlyph* glyph_cache_get(rdpGlyphCache* glyph_cache, uint8 id, uint16 index);
-FREERDP_API void glyph_cache_put(rdpGlyphCache* glyph_cache, uint8 id, uint16 index, rdpGlyph* entry);
+FREERDP_API rdpGlyph* glyph_cache_get(rdpGlyphCache* glyph_cache, uint32 id, uint32 index);
+FREERDP_API void glyph_cache_put(rdpGlyphCache* glyph_cache, uint32 id, uint32 index, rdpGlyph* entry);
 
-FREERDP_API void* glyph_cache_fragment_get(rdpGlyphCache* glyph, uint8 index, uint8* count);
-FREERDP_API void glyph_cache_fragment_put(rdpGlyphCache* glyph, uint8 index, uint8 count, void* entry);
+FREERDP_API void* glyph_cache_fragment_get(rdpGlyphCache* glyph, uint32 index, uint32* count);
+FREERDP_API void glyph_cache_fragment_put(rdpGlyphCache* glyph, uint32 index, uint32 count, void* entry);
 
 FREERDP_API void glyph_cache_register_callbacks(rdpUpdate* update);
