+++ libfreerdp-cache/pointer.c
@@ -71,7 +71,7 @@ void update_pointer_cached(rdpContext* context, POINTER_CACHED_UPDATE* pointer_c
 	Pointer_Set(context, pointer);
 }
 
-rdpPointer* pointer_cache_get(rdpPointerCache* pointer_cache, uint16 index)
+rdpPointer* pointer_cache_get(rdpPointerCache* pointer_cache, uint32 index)
 {
 	rdpPointer* pointer;
 
@@ -86,7 +86,7 @@ rdpPointer* pointer_cache_get(rdpPointerCache* pointer_cache, uint16 index)
 	return pointer;
 }
 
-void pointer_cache_put(rdpPointerCache* pointer_cache, uint16 index, rdpPointer* pointer)
+void pointer_cache_put(rdpPointerCache* pointer_cache, uint32 index, rdpPointer* pointer)
 {
 	if (index >= pointer_cache->cacheSize)
 	{
